Update warmup module
