+++ examples/fedprox_example/client.py
@@ -12,12 +12,12 @@ from torch.optim import Optimizer
 from torch.utils.data import DataLoader
 
 from examples.models.cnn_model import MnistNet
-from fl4health.client_surgery.warmed_up_module import WarmedUpModule
 from fl4health.clients.fed_prox_client import FedProxClient
 from fl4health.utils.load_data import load_mnist_data
 from fl4health.utils.metrics import Accuracy
 from fl4health.utils.random import set_all_random_seeds
 from fl4health.utils.sampler import DirichletLabelBasedSampler
+from fl4health.utils.warmed_up_module import WarmedUpModule
 
 
 class MnistFedProxClient(FedProxClient):
