+++ fl4health/clients/moon_client.py
@@ -5,11 +5,11 @@ import torch
 from flwr.common.typing import Config, NDArrays
 
 from fl4health.checkpointing.checkpointer import TorchCheckpointer
-from fl4health.client_surgery.warmed_up_module import WarmedUpModule
 from fl4health.clients.basic_client import BasicClient
 from fl4health.model_bases.moon_base import MoonModel
 from fl4health.utils.losses import Losses, LossMeterType
 from fl4health.utils.metrics import Metric
+from fl4health.utils.warmed_up_module import WarmedUpModule
 
 
 class MoonClient(BasicClient):
